moved protectedRoute file to routes folder
